+++ app/src/main/java/co/develhope/meteoapp/data/DataSource.kt
@@ -1,5 +1,7 @@
 package co.develhope.meteoapp.data
 
+import co.develhope.meteoapp.todayscreen.domain.DetailedCardForecast
+import co.develhope.meteoapp.todayscreen.domain.DomainHourlyForecast
 import org.threeten.bp.OffsetDateTime
 
 object DataSource{
@@ -74,49 +76,32 @@ object DataSource{
     fun getHomeItems()= homeItems
 
 
-    private val ForecastList: List<Forecast> = listOf(
-        TitleForecast(OffsetDateTime.now(), "Napoli", "Campania"),
-        HourlyForecastListItem(OffsetDateTime.now(), Weather.SUNNY, 20, 15,
-            DetailedCardForecast( 22, 15, 0, 6, 1)
-        ),
-        HourlyForecastListItem(OffsetDateTime.now().plusHours(1), Weather.CLOUDY, 20, 15,
-            DetailedCardForecast( 22, 15, 0, 7, 6)
-        ),
-        HourlyForecastListItem(OffsetDateTime.now().plusHours(2), Weather.RAINY, 20, 15,
-            DetailedCardForecast( 22, 15, 0, 4, 3)
-        ),
-        HourlyForecastListItem(OffsetDateTime.now().plusHours(3), Weather.NIGHT, 20, 15,
-            DetailedCardForecast( 22, 15, 0, 7, 2)
-        ),
-        HourlyForecastListItem(OffsetDateTime.now().plusHours(4), Weather.NIGHT, 20, 15,
-            DetailedCardForecast( 22, 15, 0, 7, 2)
-        ),
-        HourlyForecastListItem(OffsetDateTime.now().plusHours(5), Weather.NIGHT, 20, 15,
-            DetailedCardForecast( 22, 15, 0, 7, 2)
-        ),
-        HourlyForecastListItem(OffsetDateTime.now().plusHours(6), Weather.NIGHT, 20, 15,
-            DetailedCardForecast( 22, 15, 0, 7, 2)
-        ),
-        HourlyForecastListItem(OffsetDateTime.now().plusHours(7), Weather.NIGHT, 20, 15,
-            DetailedCardForecast( 22, 15, 0, 7, 2)
-        ),
-        HourlyForecastListItem(OffsetDateTime.now().plusHours(8), Weather.NIGHT, 20, 15,
-            DetailedCardForecast( 22, 15, 0, 7, 2)
-        ),
-        HourlyForecastListItem(OffsetDateTime.now().plusHours(9), Weather.NIGHT, 20, 15,
-            DetailedCardForecast( 22, 15, 0, 7, 2)
-        ),
-        HourlyForecastListItem(OffsetDateTime.now().plusHours(10), Weather.NIGHT, 20, 15,
-            DetailedCardForecast( 22, 15, 0, 7, 2)
-        ),
-        HourlyForecastListItem(OffsetDateTime.now().plusHours(10), Weather.NIGHT, 20, 15,
-            DetailedCardForecast( 22, 15, 0, 7, 2)
-        ),
-
-
+    private val forecast: MutableList<Forecast> = mutableListOf(
+        TitleForecast(
+            DomainHourlyForecast(
+                OffsetDateTime.now(),
+                DetailedCardForecast(
+                    perceivedTemperature = 1,
+                    coverage = 1,
+                    rain = 1,
+                    uvIndex = 1,
+                    wind = 1,
+                    wetness = 1,
+                    weather = Weather.SUNNY,
+                    celsius = 1
+                )
+            ),
+            Place(
+                name = "Naples",
+                latitude = 23.2323,
+                longitude = 23.2323
+            )
+        )
     )
 
-    fun getTodayDetailedForecast() = ForecastList
+    fun getTodayDetailedForecast(): List<Forecast>{
+        return forecast
+    }
 
     private val searchList : List<GetHourlyForecastList> = listOf(
         RecentSearches("Ricerche Recenti"),
