--- app/src/main/java/co/develhope/meteoapp/todayscreen/domain/Place.kt
@@ -1,7 +1,0 @@
-package co.develhope.meteoapp.todayscreen.domain
-
-data class Place(
-    val name: String,
-    val latitude: Double,
-    val longitude: Double
-)
